skills/logicapp-std-tool-python-port/scaffolding/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lat"
version = "0.1.0"
description = "Logic App Standard Advanced Tool — Python port"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "TODO" }]
keywords = ["azure", "logic-apps", "logic-app-standard"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "typer>=0.12.0",
    "rich>=13.7.0",
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
    "zstandard>=0.22.0",
    "azure-identity>=1.16.0",
    "azure-data-tables>=12.5.0",
    "azure-storage-blob>=12.20.0",
    "azure-storage-file-share>=12.16.0",
    "azure-storage-queue>=12.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "pytest-asyncio>=0.23.0",
    "respx>=0.21.0",          # httpx mocking
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "types-zstandard",
]

[project.scripts]
lat = "lat.cli:app"

[project.urls]
"Source (canonical)" = "https://github.com/microsoft/Logic-App-STD-Advanced-Tools"
"Source (this port)" = "TODO"

[tool.hatch.build.targets.wheel]
packages = ["src/lat"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "live: tests that require LIVE_LAT_* env vars to run against a sandbox Logic App",
]
addopts = ["-ra", "--strict-markers"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM", "PL"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
disallow_untyped_defs = true
